feat: deploy InsuranceCore to WebLogic and configure PostgreSQL datasource
- configure Wicket application and web.xml
- package insurance-web module as WAR
- align project with Java EE 8 and Wicket 9
- update Commons FileUpload dependency
- deploy application to WebLogic 14.1.1
- configure JTA datasource with PostgreSQL
- add PostgreSQL JDBC driver to WebLogic domain
- connect WebLogic to PostgreSQL Docker container
- verify Wicket, EJB, JPA and datasource integration

+++ insurance-persistence/src/main/java/com/insurance/persistence/repository/ClaimRepository.java
@@ -20,7 +20,7 @@ public class ClaimRepository {
         return selectCFromClaimC.getResultList();
     }
 
-    public Optional<Claim> findById(int id) {
+    public Optional<Claim> findById(Long id) {
         return Optional.ofNullable(entityManager.find(Claim.class, id));
     }
